+++ pkg/app/serviceauth/auth.go
@@ -1,0 +1,14 @@
+package serviceauth
+
+import (
+	"github.com/Educentr/go-project-starter-runtime/pkg/app"
+	"github.com/Educentr/go-project-starter-runtime/pkg/ds"
+)
+
+type Authorizer struct {
+	app.UnimplementedAuthorizer
+}
+
+func (a *Authorizer) GetAuthorizer() ds.Authorizer {
+	return a
+}